python/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-storage
google-cloud-bigquery
google-cloud-tasks
pymupdf
spacy
isbnlib
fastapi
uvicorn
anyio
requests


